ai-engine/requirements-rag.txt
flask
flask-cors
mysql-connector-python
numpy
faiss-cpu
